Create game.cpp
First pass at this.  Had to create some methods and variables on the fly, going back to add those in where appropriate.  Drawing not setup due to the fact that I'm not sure if I can access the underlying vector of a priority queue.  If I can, it's easier than I thought; if I can't then I have to recreate the queue each time - which isn't too terrible.